Splitted create indicator component
