+++ opencti-platform/opencti-graphql/src/modules/playbook/playbook-components.ts
@@ -17,23 +17,14 @@ import type { JSONSchemaType } from 'ajv';
 import { type BasicStoreEntityPlaybook, ENTITY_TYPE_PLAYBOOK, type PlaybookComponent } from './playbook-types';
 import { AUTOMATION_MANAGER_USER, AUTOMATION_MANAGER_USER_UUID, executionContext, isUserCanAccessStixElement, isUserInPlatformOrganization, SYSTEM_USER } from '../../utils/access';
 import { pushToConnector, pushToWorkerForConnector } from '../../database/rabbitmq';
-import { ABSTRACT_STIX_CORE_RELATIONSHIP, ABSTRACT_STIX_CYBER_OBSERVABLE, ENTITY_TYPE_CONTAINER, ENTITY_TYPE_THREAT_ACTOR } from '../../schema/general';
+import { ABSTRACT_STIX_CORE_RELATIONSHIP, ABSTRACT_STIX_CYBER_OBSERVABLE, ENTITY_TYPE_CONTAINER } from '../../schema/general';
 import type { BasicStoreRelation, StoreCommon, StoreRelation } from '../../types/store';
 import { generateInternalId, generateStandardId, idGenFromData } from '../../schema/identifier';
-import { now, observableValue, utcDate } from '../../utils/format';
+import { now, utcDate } from '../../utils/format';
 import type { StixCampaign, StixContainer, StixIncident, StixInfrastructure, StixMalware, StixReport, StixThreatActor } from '../../types/stix-2-1-sdo';
 import { convertStixToInternalTypes, generateInternalType, getParentTypes } from '../../schema/schemaUtils';
-import {
-  ENTITY_TYPE_ATTACK_PATTERN,
-  ENTITY_TYPE_CAMPAIGN,
-  ENTITY_TYPE_CONTAINER_REPORT,
-  ENTITY_TYPE_INCIDENT,
-  ENTITY_TYPE_INTRUSION_SET,
-  ENTITY_TYPE_MALWARE,
-  ENTITY_TYPE_TOOL,
-  isStixDomainObjectContainer,
-} from '../../schema/stixDomainObject';
-import type { CyberObjectExtension, StixBundle, StixCoreObject, StixCyberObject, StixObject, StixOpenctiExtension } from '../../types/stix-2-1-common';
+import { ENTITY_TYPE_CONTAINER_REPORT, isStixDomainObjectContainer } from '../../schema/stixDomainObject';
+import type { CyberObjectExtension, StixBundle, StixCyberObject, StixObject, StixOpenctiExtension } from '../../types/stix-2-1-common';
 import { STIX_EXT_OCTI, STIX_EXT_OCTI_SCO } from '../../types/stix-2-1-extensions';
 import { connectorsForPlaybook } from '../../database/repository';
 import { fullEntitiesList, fullRelationsList, storeLoadById } from '../../database/middleware-loader';
@@ -46,16 +37,13 @@ import { convertToNotificationUser, type DigestEvent, EVENT_NOTIFICATION_VERSION
 import { storeNotificationEvent } from '../../database/stream/stream-handler';
 import { ENTITY_TYPE_SETTINGS } from '../../schema/internalObject';
 import { isStixCyberObservable } from '../../schema/stixCyberObservable';
-import { createStixPattern } from '../../python/pythonBridge';
-import { generateKeyValueForIndicator } from '../../domain/stixCyberObservable';
-import { RELATION_BASED_ON, RELATION_INDICATES } from '../../schema/stixCoreRelationship';
+import { RELATION_BASED_ON } from '../../schema/stixCoreRelationship';
 import type { StixRelation } from '../../types/stix-2-1-sro';
-import { extractValidObservablesFromIndicatorPattern, STIX_PATTERN_TYPE } from '../../utils/syntax';
+import { extractValidObservablesFromIndicatorPattern } from '../../utils/syntax';
 import { isStixMatchFilterGroup } from '../../utils/filtering/filtering-stix/stix-filtering';
 import { ENTITY_TYPE_INDICATOR, type StixIndicator } from '../indicator/indicator-types';
 import { ENTITY_TYPE_CONTAINER_TASK, type StixTask, type StoreEntityTask } from '../task/task-types';
 import { FilterMode } from '../../generated/graphql';
-import { schemaTypesDefinition } from '../../schema/schema-types';
 import { generateCreateMessage } from '../../database/data-changes';
 import { findAllByCaseTemplateId } from '../task/task-domain';
 import type { BasicStoreEntityTaskTemplate } from '../task/task-template/task-template-types';
@@ -72,6 +60,7 @@ import { PLAYBOOK_SHARING_COMPONENT } from './components/sharing-component';
 import { PLAYBOOK_UNSHARING_COMPONENT } from './components/unsharing-component';
 import { PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT } from './components/access-restrictions-component';
 import { PLAYBOOK_REMOVE_ACCESS_RESTRICTIONS_COMPONENT } from './components/remove-access-restrictions-component';
+import { PLAYBOOK_CREATE_INDICATOR_COMPONENT } from './components/create-indicator-component';
 
 // region built in playbook components
 interface LoggerConfiguration {
@@ -698,237 +687,7 @@ const PLAYBOOK_NOTIFIER_COMPONENT: PlaybookComponent<NotifierConfiguration> = {
     return { output_port: undefined, bundle };
   },
 };
-interface CreateIndicatorConfiguration {
-  all: boolean;
-  wrap_in_container: boolean;
-  types: string[];
-}
-const PLAYBOOK_CREATE_INDICATOR_COMPONENT_SCHEMA: JSONSchemaType<CreateIndicatorConfiguration> = {
-  type: 'object',
-  properties: {
-    types: {
-      type: 'array',
-      default: [],
-      $ref: 'Types',
-      items: { type: 'string', oneOf: [] },
-    },
-    all: { type: 'boolean', $ref: 'Create indicators from all observables in the bundle', default: false },
-    wrap_in_container: { type: 'boolean', $ref: 'If main entity is a container, wrap indicators in container', default: false },
-  },
-  required: [],
-};
-const PLAYBOOK_CREATE_INDICATOR_COMPONENT: PlaybookComponent<CreateIndicatorConfiguration> = {
-  id: 'PLAYBOOK_CREATE_INDICATOR_COMPONENT',
-  name: 'Promote observable to indicator',
-  description: 'Create an indicator based on an observable',
-  icon: 'indicator',
-  is_entry_point: false,
-  is_internal: true,
-  ports: [{ id: 'out', type: 'out' }, { id: 'unmodified', type: 'out' }],
-  configuration_schema: PLAYBOOK_CREATE_INDICATOR_COMPONENT_SCHEMA,
-  schema: async () => {
-    const types = schemaTypesDefinition.get(ABSTRACT_STIX_CYBER_OBSERVABLE);
-    const elements = types.map((t) => ({ const: t, title: t }))
-      .sort((a, b) => (a.title.toLowerCase() > b.title.toLowerCase() ? 1 : -1));
-    const schemaElement = { properties: { types: { items: { oneOf: elements } } } };
-    return R.mergeDeepRight<JSONSchemaType<CreateIndicatorConfiguration>, any>(PLAYBOOK_CREATE_INDICATOR_COMPONENT_SCHEMA, schemaElement);
-  },
-  executor: async ({ playbookNode, dataInstanceId, bundle }) => {
-    const { all, wrap_in_container, types } = playbookNode.configuration;
-    const context = executionContext('playbook_components');
-    const baseData = extractBundleBaseElement(dataInstanceId, bundle);
-    const observables = [baseData];
-    if (all) {
-      pushAll(observables, bundle.objects);
-    }
-    const { type: baseDataType, id } = baseData.extensions[STIX_EXT_OCTI];
-    const isBaseDataAContainer = isStixDomainObjectContainer(baseDataType);
-    const objectsToPush: StixObject[] = [];
-    for (let index = 0; index < observables.length; index += 1) {
-      const observable = observables[index] as StixCyberObject;
-      let { type } = observable.extensions[STIX_EXT_OCTI];
-      if (isStixCyberObservable(type) && (isEmptyField(types) || types.includes(type))) {
-        const indicatorName = observableValue({ ...observable, entity_type: type });
-        const { key, value } = generateKeyValueForIndicator(type, indicatorName, observable);
-        if (key.includes('Artifact')) {
-          type = 'StixFile';
-        }
-        const pattern = await createStixPattern(context, AUTOMATION_MANAGER_USER, key, value);
-        const score = observable.x_opencti_score ?? observable.extensions[STIX_EXT_OCTI_SCO]?.score;
-        const { granted_refs } = observable.extensions[STIX_EXT_OCTI];
-        if (pattern) {
-          const indicatorData = {
-            name: indicatorName,
-            x_opencti_main_observable_type: type,
-            x_opencti_score: score,
-            pattern,
-            pattern_type: STIX_PATTERN_TYPE,
-            extensions: {
-              [STIX_EXT_OCTI]: {
-                extension_type: 'property-extension',
-                type: ENTITY_TYPE_INDICATOR,
-                main_observable_type: type,
-                score,
-              },
-            },
-          };
-          const indicatorStandardId = generateStandardId(ENTITY_TYPE_INDICATOR, indicatorData);
-          const storeIndicator = {
-            internal_id: generateInternalId(),
-            standard_id: indicatorStandardId,
-            entity_type: ENTITY_TYPE_INDICATOR,
-            parent_types: getParentTypes(ENTITY_TYPE_INDICATOR),
-            ...indicatorData,
-          } as StoreCommon;
-          const indicator = convertStoreToStix_2_1(storeIndicator) as StixIndicator;
-          if (observable.object_marking_refs) {
-            indicator.object_marking_refs = observable.object_marking_refs;
-          }
-          if (observable.extensions[STIX_EXT_OCTI_SCO]?.labels) {
-            indicator.labels = observable.extensions[STIX_EXT_OCTI_SCO].labels;
-          }
-          if (observable.extensions[STIX_EXT_OCTI_SCO]?.created_by_ref) {
-            indicator.created_by_ref = observable.extensions[STIX_EXT_OCTI_SCO].created_by_ref;
-          }
-          if (observable.extensions[STIX_EXT_OCTI_SCO]?.external_references) {
-            indicator.external_references = observable.extensions[STIX_EXT_OCTI_SCO].external_references;
-          }
-          if (granted_refs) {
-            indicator.extensions[STIX_EXT_OCTI].granted_refs = granted_refs;
-          }
-          objectsToPush.push(indicator);
-          if (wrap_in_container && isBaseDataAContainer) {
-            (baseData as StixContainer).object_refs.push(indicator.id);
-          }
-          const relationBaseData = {
-            source_ref: indicator.id,
-            target_ref: observable.id,
-            relationship_type: RELATION_BASED_ON,
-          };
-          const relationStandardId = idGenFromData('relationship', relationBaseData);
-          const relationship = {
-            id: relationStandardId,
-            type: 'relationship',
-            ...relationBaseData,
-            object_marking_refs: observable.object_marking_refs ?? [],
-            created: now(),
-            modified: now(),
-            extensions: {
-              [STIX_EXT_OCTI]: {
-                extension_type: 'property-extension',
-                type: RELATION_BASED_ON,
-              },
-            },
-          } as StixRelation;
-          if (granted_refs) {
-            relationship.extensions[STIX_EXT_OCTI].granted_refs = granted_refs;
-          }
-          objectsToPush.push(relationship);
 
-          // Resolve relationships in the bundle
-          const stixRelationshipsInBundle = bundle.objects.filter((r) => r.type === 'relationship') as StixRelation[];
-          const stixRelationships = stixRelationshipsInBundle.filter((r) => r.relationship_type === 'related-to'
-            && r.source_ref === baseData.id
-            && (
-              r.target_ref.startsWith('threat-actor')
-              || r.target_ref.startsWith('intrusion-set')
-              || r.target_ref.startsWith('campaign')
-              || r.target_ref.startsWith('malware')
-              || r.target_ref.startsWith('incident')
-              || r.target_ref.startsWith('tool')
-              || r.target_ref.startsWith('attack-pattern')
-            ));
-          for (let indexStixRelationships = 0; indexStixRelationships < stixRelationships.length; indexStixRelationships += 1) {
-            const stixRelationship = stixRelationships[indexStixRelationships] as StixRelation;
-            const relationIndicatesBaseData = {
-              source_ref: indicator.id,
-              target_ref: stixRelationship.target_ref,
-              relationship_type: RELATION_INDICATES,
-            };
-            const relationIndicatesStandardId = idGenFromData('relationship', relationIndicatesBaseData);
-            const relationshipIndicates = {
-              id: relationIndicatesStandardId,
-              type: 'relationship',
-              ...relationIndicatesBaseData,
-              object_marking_refs: observable.object_marking_refs ?? [],
-              created: now(),
-              modified: now(),
-              extensions: {
-                [STIX_EXT_OCTI]: {
-                  extension_type: 'property-extension',
-                  type: RELATION_INDICATES,
-                },
-              },
-            } as StixRelation;
-            if (granted_refs) {
-              relationshipIndicates.extensions[STIX_EXT_OCTI].granted_refs = granted_refs;
-            }
-            objectsToPush.push(relationshipIndicates);
-          }
-          // Resolve relationships in database
-          if (isNotEmptyField(id)) {
-            const relationsOfObservables = await fullRelationsList(
-              context,
-              AUTOMATION_MANAGER_USER,
-              ABSTRACT_STIX_CORE_RELATIONSHIP,
-              {
-                fromOrToId: id,
-                toTypes: [
-                  ENTITY_TYPE_THREAT_ACTOR,
-                  ENTITY_TYPE_INTRUSION_SET,
-                  ENTITY_TYPE_CAMPAIGN,
-                  ENTITY_TYPE_MALWARE,
-                  ENTITY_TYPE_INCIDENT,
-                  ENTITY_TYPE_TOOL,
-                  ENTITY_TYPE_ATTACK_PATTERN,
-                ],
-                baseData: true,
-                indices: READ_RELATIONSHIPS_INDICES,
-              },
-            ) as StoreRelation[];
-            const idsToResolve = R.uniq(relationsOfObservables.map((r) => r.toId));
-            const elements = await stixLoadByIds(context, AUTOMATION_MANAGER_USER, idsToResolve);
-            for (let indexElements = 0; indexElements < elements.length; indexElements += 1) {
-              const element = elements[indexElements] as StixCoreObject;
-              const relationIndicatesBaseData = {
-                source_ref: indicator.id,
-                target_ref: element.id,
-                relationship_type: RELATION_INDICATES,
-              };
-              const relationIndicatesStandardId = idGenFromData('relationship', relationIndicatesBaseData);
-              const relationshipIndicates = {
-                id: relationIndicatesStandardId,
-                type: 'relationship',
-                ...relationIndicatesBaseData,
-                object_marking_refs: observable.object_marking_refs ?? [],
-                created: now(),
-                modified: now(),
-                extensions: {
-                  [STIX_EXT_OCTI]: {
-                    extension_type: 'property-extension',
-                    type: RELATION_INDICATES,
-                  },
-                },
-              } as StixRelation;
-              if (granted_refs) {
-                relationshipIndicates.extensions[STIX_EXT_OCTI].granted_refs = granted_refs;
-              }
-              objectsToPush.push(relationshipIndicates);
-            }
-          }
-          if (wrap_in_container && isBaseDataAContainer) {
-            (baseData as StixContainer).object_refs.push(relationship.id);
-          }
-        }
-      }
-    }
-    if (objectsToPush.length > 0) {
-      pushAll(bundle.objects, objectsToPush);
-      return { output_port: 'out', bundle: { ...bundle, objects: bundle.objects.map((n) => (n.id === baseData.id ? baseData : n)) } };
-    }
-    return { output_port: 'unmodified', bundle };
-  },
-};
 interface CreateObservableConfiguration {
   all: boolean;
   wrap_in_container: boolean;
